+++ runtime/mirror/array-inl.h
@@ -130,7 +130,7 @@ class SetLengthToUsableSizeVisitor {
     // Avoid AsArray as object is not yet in live bitmap or allocation stack.
     Array* array = down_cast<Array*>(obj);
     // DCHECK(array->IsArrayInstance());
-    int32_t length = (usable_size - header_size_) >> component_size_shift_;
+    int32_t length = static_cast<int32_t>((usable_size - header_size_) >> component_size_shift_);
     DCHECK_GE(length, minimum_length_);
     uint8_t* old_end = reinterpret_cast<uint8_t*>(array->GetRawData(1U << component_size_shift_,
                                                                     minimum_length_));
@@ -203,7 +203,7 @@ inline void PrimitiveArray<T>::VisitRoots(RootVisitor* visitor) {
 }
 
 template<typename T>
-inline PrimitiveArray<T>* PrimitiveArray<T>::Alloc(Thread* self, size_t length) {
+inline PrimitiveArray<T>* PrimitiveArray<T>::Alloc(Thread* self, int32_t length) {
   Array* raw_array = Array::Alloc<true>(self, GetArrayClass(), length,
                                         ComponentSizeShiftWidth(sizeof(T)),
                                         Runtime::Current()->GetHeap()->GetCurrentAllocator());
