+++ runtime/utf.cc
@@ -115,7 +115,7 @@ int32_t ComputeUtf16Hash(const uint16_t* chars, size_t char_count) {
   return static_cast<int32_t>(hash);
 }
 
-size_t ComputeModifiedUtf8Hash(const char* chars) {
+int32_t ComputeModifiedUtf8Hash(const char* chars) {
   size_t hash = 0;
   while (*chars != '\0') {
     hash = hash * 31 + *chars++;
